feature: Rename TeamRecruitment response classes to Recruitment
