+++ data/src/main/java/in/koreatech/koin/data/response/recruitment/RecruitmentListResponse.kt
@@ -2,9 +2,9 @@ package `in`.koreatech.koin.data.response.recruitment
 
 import com.google.gson.annotations.SerializedName
 
-data class TeamRecruitmentListResponse(
+data class RecruitmentListResponse(
     @SerializedName("recruitments")
-    val recruitments: List<TeamRecruitmentResponse>,
+    val recruitments: List<RecruitmentResponse>,
     @SerializedName("total_count")
     val totalCount: Long,
     @SerializedName("current_count")
